# docs/swagger.yaml
basePath: /api/v1
definitions:
  handlers.ApproveRequest:
    properties:
      contract_no:
        type: string
      remark:
        type: string
    type: object
  handlers.ChangeOfficerRequest:
    properties:
      officer_id:
        type: integer
      remark:
        type: string
    type: object
...
  handlers.CreateApptRequest:
    properties:
      appt_date:
        type: string
      appt_time:
        type: string
      loan_appt_id:
        type: integer
      location:
        type: string
      remark:
        type: string
    type: object
  handlers.CreateMortgageRequest:
    properties:
      amount:
        type: number
      collateral:
        type: string
      guarantor_memb_no:
        type: string
      loan_type_id:
        type: integer
      memb_no:
        type: string
      purpose:
        type: string
      remark:
        type: string
    type: object
...
  handlers.RejectRequest:
    properties:
      remark:
        type: string
    type: object
...
  handlers.UpdateDocRequest:
    properties:
      doc_id:
        type: integer
      is_submitted:
        type: boolean
      remark:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:3000
info:
  contact:
    email: support@spsc.or.th
    name: API Support
  description: ระบบสินเชื่อสหกรณ์ SPSC loanEasy v1.0 API
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: SPSC loanEasy API
  version: "1.0"
paths:
...
  /api/v1/mortgages:
    get:
      consumes:
      - application/json
      description: List all mortgages (Officer only)
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Items per page
        in: query
        name: limit
        type: integer
      - description: Filter by officer ID
        in: query
        name: officer_id
        type: integer
      - description: Filter by step ID
        in: query
        name: step_id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: List mortgages
      tags:
      - Mortgages
    post:
      consumes:
      - application/json
      description: Create a new mortgage (Officer only)
      parameters:
      - description: Mortgage data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateMortgageRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create mortgage
      tags:
      - Mortgages
  /api/v1/mortgages/{id}:
    get:
      consumes:
      - application/json
      description: Get a specific mortgage (Officer only)
      parameters:
      - description: Mortgage ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get mortgage by ID
      tags:
      - Mortgages
  /api/v1/mortgages/{id}/approve:
    put:
      consumes:
      - application/json
      description: Approve a mortgage (Officer only)
      parameters:
      - description: Mortgage ID
        in: path
        name: id
        required: true
        type: integer
      - description: Approve data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ApproveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Approve mortgage
      tags:
      - Mortgages
  /api/v1/mortgages/{id}/appts:
    get:
      consumes:
      - application/json
      description: Get mortgage appointments
      parameters:
      - description: Mortgage ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get appointments
      tags:
      - Mortgages
    post:
      consumes:
      - application/json
      description: Create a new appointment (Officer only)
      parameters:
      - description: Mortgage ID
        in: path
        name: id
        required: true
        type: integer
      - description: Appointment data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateApptRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Create appointment
      tags:
      - Mortgages
  /api/v1/mortgages/{id}/appts/{appt_id}/complete:
    put:
      consumes:
      - application/json
      description: Mark appointment as completed (Officer only)
      parameters:
      - description: Mortgage ID
        in: path
        name: id
        required: true
        type: integer
      - description: Appointment ID
        in: path
        name: appt_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Complete appointment
      tags:
      - Mortgages
  /api/v1/mortgages/{id}/docs:
    get:
      consumes:
      - application/json
      description: Get mortgage document checklist
      parameters:
      - description: Mortgage ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get mortgage documents
      tags:
      - Mortgages
    put:
      consumes:
      - application/json
      description: Update mortgage document status (Officer only)
      parameters:
      - description: Mortgage ID
        in: path
        name: id
        required: true
        type: integer
      - description: Document data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.UpdateDocRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Update document status
      tags:
      - Mortgages
  /api/v1/mortgages/{id}/history:
    get:
      consumes:
      - application/json
      description: Get mortgage transaction history
      parameters:
      - description: Mortgage ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get mortgage history
      tags:
      - Mortgages
  /api/v1/mortgages/{id}/officer:
    put:
      consumes:
      - application/json
      description: Change the responsible officer (Admin only)
      parameters:
      - description: Mortgage ID
        in: path
        name: id
        required: true
        type: integer
      - description: Officer data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ChangeOfficerRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Change officer
      tags:
      - Mortgages
  /api/v1/mortgages/{id}/reject:
    put:
      consumes:
      - application/json
      description: Reject a mortgage (Officer only)
      parameters:
      - description: Mortgage ID
        in: path
        name: id
        required: true
        type: integer
      - description: Reject data
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.RejectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/response.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Reject mortgage
      tags:
      - Mortgages
...
  /api/v1/mortgages/my:
    get:
      consumes:
      - application/json
      description: Get current user's mortgages
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Get my mortgages
      tags:
      - Mortgages
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"